tests/submission/PhilipTaylor/canvas/2d.text.draw.align.end.rtl.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: 2d.text.draw.align.end.rtl</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
<style>
@font-face {
  font-family: CanvasTest;
  src: url("../fonts/CanvasTest.ttf");
}
</style>
</head><body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.text.html">text</a>.<a href="index.2d.text.draw.html">draw</a>.<a href="index.2d.text.draw.align.html">align</a>.<a href="index.2d.text.draw.align.end.html">end</a>.rtl</h1>
<p class="desc">textAlign end with rtl is the left edge</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.text.align.left">2d.text.align.left</a>
</li><li><a href="../annotated-spec/canvas.html#testrefs.2d.text.draw.direction">2d.text.draw.direction</a>
</li></ul>
</div>

<span style="font-family: CanvasTest; position: absolute; visibility: hidden">A</span>
<p class="output">Actual output:</p>
<canvas class="output" dir="rtl" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:</p><p><img alt="" class="output expected" id="expected" src="green-100x50.png">
</p><ul id="d"></ul>
<script>
var t = async_test("textAlign end with rtl is the left edge");
_addTest(function(canvas, ctx) {

ctx.font = '50px CanvasTest';
deferTest();
setTimeout(wrapFunction(function () {
    ctx.fillStyle = '#f00';
    ctx.fillRect(0, 0, 100, 50);
    ctx.fillStyle = '#0f0';
    ctx.textAlign = 'end';
    ctx.fillText('DD', 0, 37.5);
    _assertPixelApprox(canvas, 5,5, 0,255,0,255, "5,5", "0,255,0,255", 2);
    _assertPixelApprox(canvas, 95,5, 0,255,0,255, "95,5", "0,255,0,255", 2);
    _assertPixelApprox(canvas, 25,25, 0,255,0,255, "25,25", "0,255,0,255", 2);
    _assertPixelApprox(canvas, 75,25, 0,255,0,255, "75,25", "0,255,0,255", 2);
    _assertPixelApprox(canvas, 5,45, 0,255,0,255, "5,45", "0,255,0,255", 2);
    _assertPixelApprox(canvas, 95,45, 0,255,0,255, "95,45", "0,255,0,255", 2);
}), 500);


});
</script>

<div id="log"></div>
</body></html>
